ILUSNITOM, Tabula Rasa
In the beginning, there was little but a few dwarves who beseeched Armok to forge a world anew upon his anvil of great hewing. And it was done, and it was so. And thus was birthed Ilusnitom, which to all the peoples of the world, is known as The Everseeing Planet.